fix: [RABBIT-147] 펀딩 성공/실패 모달 수정
